+++ src/pages/help-centre/index.js
@@ -1,0 +1,208 @@
+import React, { Component } from 'react'
+import matchSorter from 'match-sorter'
+import styled from 'styled-components'
+import { navigate } from '@reach/router'
+import SEO from '../../components/containers/seo'
+import Layout from '../../components/layout/layout'
+import Container from '../../components/containers/container'
+import { localize, WithIntl } from '../../components/localization'
+import { getLocationHash } from '../../common/utility'
+import { articles } from './_help-articles'
+import { ArticleSection } from './_article-section'
+import { SearchSuccess, SearchError } from './_search-results'
+
+import SearchIcon from 'images/svg/search.svg'
+import CrossIcon from 'images/svg/cross.svg'
+
+const getAllArticles = articles =>
+    articles
+        .map(category => category.articles)
+        // flatten the array, gatsby build does not support .flat() yet
+        .reduce((arr, article_arr) => arr.concat(article_arr), [])
+
+const SearchSection = styled.section`
+    ${Backdrop} {
+        max-height: ${props => (props.show ? '100rem' : '0')};
+        transition: ${props =>
+            props.has_transition ? 'max-height 0.6s ease-in-out' : 'none'};
+        overflow: hidden;
+    }
+`
+const Backdrop = styled.div`
+    background-color: var(--color-black);
+    padding: 12rem 0;
+`
+
+const SearchIconBig = styled(SearchIcon)`
+    position: absolute;
+    left: 0;
+    top: 3px;
+`
+
+const SearchCrossIcon = styled(CrossIcon)`
+    :hover {
+        cursor: pointer;
+    }
+`
+
+const SearchForm = styled.form`
+    position: relative;
+    padding-left: 6.4rem;
+`
+
+const Search = styled.input`
+    font-size: 4.4rem;
+    font-weight: bold;
+    color: var(--color-white);
+    background-color: var(--color-black);
+    border: none;
+    outline: none;
+
+    ::placeholder {
+        color: var(--color-grey-3);
+    }
+`
+
+const ResultWrapper = styled.div`
+    margin-top: 4rem;
+
+    > :first-child {
+        margin-bottom: 1.6rem;
+    }
+`
+
+class HelpCentre extends Component {
+    constructor(props) {
+        super(props)
+        const all_articles = getAllArticles(articles)
+
+        this.state = {
+            all_articles,
+            search: '',
+            selected_article: null,
+            search_has_transition: false,
+            toggle_search: true,
+        }
+    }
+
+    handleInputChange = e => {
+        e.preventDefault()
+        const { name, value } = e.target
+        this.setState({ [name]: `${value.trim()}` })
+    }
+
+    handleSubmit = e => e.preventDefault()
+
+    handleSelectArticle = article => {
+        navigate(`#${article.label}`)
+        this.setState({
+            selected_article: article,
+            toggle_search: false,
+            search_has_transition: false,
+            search: '',
+        })
+    }
+
+    toggleSearch = () =>
+        this.setState(prevState => ({
+            toggle_search: !prevState.toggle_search,
+            search_has_transition: true,
+        }))
+
+    clearSearch = () => this.setState({ search: '' })
+
+    componentDidMount = () => {
+        const current_label = getLocationHash()
+        if (current_label) {
+            const selected_article = this.state.all_articles.find(
+                article => article.label === current_label,
+            )
+            this.setState({
+                selected_article,
+                toggle_search: false,
+                search_has_transition: false,
+            })
+        }
+    }
+
+    componentDidUpdate = () => {
+        const current_label = getLocationHash()
+
+        if (!current_label && this.state.selected_article) {
+            this.setState({
+                selected_article: null,
+                toggle_search: true,
+                search_has_transition: false,
+            })
+        }
+    }
+
+    render() {
+        const {
+            all_articles,
+            search,
+            selected_article,
+            toggle_search,
+            search_has_transition,
+        } = this.state
+
+        const filtered_articles = matchSorter(all_articles, search, {
+            keys: ['title', 'sub_category'],
+        })
+        const has_results = !!filtered_articles.length
+        return (
+            <Layout>
+                <SEO title={localize('Help centre')} />
+                <SearchSection
+                    show={toggle_search}
+                    has_transition={search_has_transition}
+                >
+                    <Backdrop>
+                        <Container align="normal" direction="column">
+                            <SearchForm
+                                onSubmit={this.handleSubmit}
+                                autoComplete="off"
+                            >
+                                <SearchIconBig />
+                                <Search
+                                    autoFocus
+                                    name="search"
+                                    value={search}
+                                    onChange={this.handleInputChange}
+                                    placeholder={localize('How can we help?')}
+                                    data-lpignore="true"
+                                />
+                                {search.length && (
+                                    <SearchCrossIcon
+                                        onClick={this.clearSearch}
+                                    />
+                                )}
+                            </SearchForm>
+                            <ResultWrapper>
+                                {has_results && search.length && (
+                                    <SearchSuccess
+                                        suggested_topics={filtered_articles}
+                                        onClick={this.handleSelectArticle}
+                                        max_length={3}
+                                    />
+                                )}
+                                {!has_results && (
+                                    <SearchError search={search} />
+                                )}
+                            </ResultWrapper>
+                        </Container>
+                    </Backdrop>
+                </SearchSection>
+                <ArticleSection
+                    articles={articles}
+                    all_articles={all_articles}
+                    selected_article={selected_article}
+                    handleSelectArticle={this.handleSelectArticle}
+                    toggleSearch={this.toggleSearch}
+                />
+            </Layout>
+        )
+    }
+}
+
+export default WithIntl()(HelpCentre)
